Modified logic to account for attribute change.
SportsBlog changed their XML/HTML such that the first paragraph can be
contained in a DIV with the CLASS attribute containing "text
bigStartParagraph". What users see are BLOG posts that didn't have any
text underneath the title.

I modified the logic to check if any portion of the CLASS attribute
contained 'text'. If it did, then the text is captured as before.