Switched to sportsblog.com
My sports BLOG is now on sportsblog.com, but in the past month or so
they stopped providing an RSS feed! So now I have to make my own. This
is the first bit of code that accesses sportsblog.com, and walks its
HTML for the parts that I need.