Added link hash to store links
Since the URLs aren't IDs, I'll need to store the original link.

I pondered the design of using hashes. Since the ID for the article is
now just a counter, why in the world was I using hashes? It was because
the old URLs were:

    blogpost/12345
    blogpost/39890
    blogpost/55810
    etc.

These IDs (12345, 39890, 55810) were the keys for the hashes. Sorting
them represented the order of the articles! All that goes out the door
since the IDs are now sequential.