Fixes issue with blank preview text
Due to layout and HTML changes, the original code had problems
retrieving the initial text. This loop ensures that it picks up
the text by walking through the entire HTML.

(There is I think a danger with an infinite loop, if the HTML doesn't
have _any_ text, but that's a low possibility. I also do not run this
on a hosted server, so I should be able to catch it if that happens.)