Revamp SportsBlog feed generation
Lots of changes here, all due to a redesign on SportsBlog. The main
changes:

1) Retrieves list of articles using <h3> element instead of <article>
element

2) No longer parses byline, as this does't exist anymore

3) Retrieves article subject and link in one function call:
get_subject_and_link(). This replaces get_subject().

4) Adds some logic in get_entry() to handle <div> elements that have an
'article-content' id. These elements somehow have the article text
inside of JSON in the corresponding data-article-content attribute. This
logic brings in JSON::XS to parse the JSON.

5) Updates get_pubdate() to reflect new layout.

6) Updates version of the code.

This was a lengthy change, and it made me want to rewrite the whole
thing in Python. But in the end, I was able to see something work. As
with a036110d5d, I no longer run this on a server, but rather on my
local laptop, so I'll see any errors right away.